pyproject.toml
[project]
name = "semantic-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "promptflow-azure>=1.18.0",
    "promptflow>=1.18.0",
    "python-dotenv>=1.1.0",
    "semantic-kernel>=1.32.1",
    "azure-search-documents>=11.5.2",
]
